Fixed issue whereby VS kept toggling between SubType Component and User Control for some custom controls.
In the case of DoubleBufferedPanel I removed the Designer attributes from
the class definition.
In the case of the EnableableComboBoxEditingControl I moved it to a new
file because the other classes it was sharing the old file with did not
derive from Component.
