+++ XenAdmin/Controls/ComboBoxes/EnableableComboBoxEditingControl.cs
@@ -1,0 +1,103 @@
+﻿/* Copyright (c) Citrix Systems, Inc. 
+ * All rights reserved. 
+ * 
+ * Redistribution and use in source and binary forms, 
+ * with or without modification, are permitted provided 
+ * that the following conditions are met: 
+ * 
+ * *   Redistributions of source code must retain the above 
+ *     copyright notice, this list of conditions and the 
+ *     following disclaimer. 
+ * *   Redistributions in binary form must reproduce the above 
+ *     copyright notice, this list of conditions and the 
+ *     following disclaimer in the documentation and/or other 
+ *     materials provided with the distribution. 
+ * 
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND 
+ * CONTRIBUTORS "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, 
+ * INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF 
+ * MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE 
+ * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR 
+ * CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, 
+ * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, 
+ * BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR 
+ * SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS 
+ * INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, 
+ * WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING 
+ * NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE 
+ * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF 
+ * SUCH DAMAGE.
+ */
+
+using System;
+using System.Windows.Forms;
+
+
+namespace XenAdmin.Controls
+{
+    
+    /// <summary>
+    /// The class providing the proxy for the hosted item in the combobox cell
+    /// </summary>
+    public class EnableableComboBoxEditingControl : EnableableComboBox, IDataGridViewEditingControl
+    {
+        public object EditingControlFormattedValue
+        {
+            get { return SelectedItem; }
+            set { SelectedItem = value; }
+        }
+
+        public object GetEditingControlFormattedValue(DataGridViewDataErrorContexts context)
+        {
+           return EditingControlFormattedValue;
+        }
+
+        public void ApplyCellStyleToEditingControl(DataGridViewCellStyle dataGridViewCellStyle)
+        {
+            Font = dataGridViewCellStyle.Font;
+        }
+
+        public int EditingControlRowIndex { get; set; }
+
+        public bool EditingControlWantsInputKey(Keys key, bool dataGridViewWantsInputKey)
+        {
+            switch (key & Keys.KeyCode)
+            {
+                case Keys.Left:
+                case Keys.Up:
+                case Keys.Down:
+                case Keys.Right:
+                case Keys.Home:
+                case Keys.End:
+                case Keys.PageDown:
+                case Keys.PageUp:
+                    return true;
+                default:
+                    return !dataGridViewWantsInputKey;
+            }
+        }
+
+        public void PrepareEditingControlForEdit(bool selectAll){}
+
+        public bool RepositionEditingControlOnValueChange
+        {
+            get { return false; }
+        }
+
+        public DataGridView EditingControlDataGridView { get; set; }
+
+        public bool EditingControlValueChanged { get; set; }
+
+        public Cursor EditingPanelCursor
+        {
+            get { return base.Cursor; }
+        }
+
+        protected override void OnSelectedValueChanged(EventArgs e)
+        {
+            EditingControlValueChanged = true;
+            EditingControlDataGridView.NotifyCurrentCellDirty(true);
+            base.OnSelectedValueChanged(e);
+        }
+    }
+}
