+++ XenAdmin/Controls/DataGridViewEnableableComboBoxCell.cs
@@ -110,77 +110,6 @@ namespace XenAdmin.Controls
     }
 
     /// <summary>
-    /// The class providing the proxy for the hosted item in the combobox cell
-    /// </summary>
-    public class EnableableComboBoxEditingControl : EnableableComboBox, IDataGridViewEditingControl
-    {
-        private bool valueChanged;
-
-        public object EditingControlFormattedValue
-        {
-            get { return SelectedItem; }
-            set { SelectedItem = value; }
-        }
-
-        public object GetEditingControlFormattedValue(DataGridViewDataErrorContexts context)
-        {
-           return EditingControlFormattedValue;
-        }
-
-        public void ApplyCellStyleToEditingControl(DataGridViewCellStyle dataGridViewCellStyle)
-        {
-            Font = dataGridViewCellStyle.Font;
-        }
-
-        public int EditingControlRowIndex { get; set; }
-
-        public bool EditingControlWantsInputKey(Keys key, bool dataGridViewWantsInputKey)
-        {
-            switch (key & Keys.KeyCode)
-            {
-                case Keys.Left:
-                case Keys.Up:
-                case Keys.Down:
-                case Keys.Right:
-                case Keys.Home:
-                case Keys.End:
-                case Keys.PageDown:
-                case Keys.PageUp:
-                    return true;
-                default:
-                    return !dataGridViewWantsInputKey;
-            }
-        }
-
-        public void PrepareEditingControlForEdit(bool selectAll){}
-
-        public bool RepositionEditingControlOnValueChange
-        {
-            get { return false; }
-        }
-
-        public DataGridView EditingControlDataGridView { get; set; }
-
-        public bool EditingControlValueChanged
-        {
-            get { return valueChanged; }
-            set { valueChanged = value; }
-        }
-
-        public Cursor EditingPanelCursor
-        {
-            get { return base.Cursor; }
-        }
-
-        protected override void OnSelectedValueChanged(EventArgs e)
-        {
-            valueChanged = true;
-            EditingControlDataGridView.NotifyCurrentCellDirty(true);
-            base.OnSelectedValueChanged(e);
-        }
-    }
-
-    /// <summary>
     /// A column to host the bespoke DataGridViewEnableableComboBoxCell in the DataGridView
     /// </summary>
     public class EnableableComboBoxColumn : DataGridViewColumn
