+++ XenAdmin/XenAdmin.csproj
@@ -165,6 +165,9 @@
       <DependentUpon>DataGridViewEx.cs</DependentUpon>
     </Compile>
     <Compile Include="Controls\DataGridViewTextAndImageCell.cs" />
+    <Compile Include="Controls\ComboBoxes\EnableableComboBoxEditingControl.cs">
+      <SubType>Component</SubType>
+    </Compile>
     <Compile Include="Controls\FolderChangeDialogTreeView.cs">
       <SubType>Component</SubType>
     </Compile>
